updated for latest 1.4_mergedtesting commit
Commit: https://github.com/tModLoader/tModLoader/commit/4745012b5b35cbb60412ab6701f841ac8c524da5
