+++ CheatSheet.cs
@@ -12,12 +12,10 @@ using System.Net;
 using System.Reflection;
 using Terraria;
 using Terraria.Chat;
-using Terraria.GameContent;
 using Terraria.ID;
 using Terraria.Localization;
 using Terraria.ModLoader;
 using Terraria.UI;
-using Terraria.UI.Chat;
 
 // TODO: move windows below inventory
 // TODO: Filter recipes with unobtainables.
@@ -28,7 +26,7 @@ using Terraria.UI.Chat;
 
 namespace CheatSheet
 {
-	internal class CheatSheet : Mod
+    internal class CheatSheet : Mod
 	{
 		internal static ModHotKey ToggleCheatSheetHotbarHotKey;
 		internal static CheatSheet instance;
@@ -345,98 +343,10 @@ namespace CheatSheet
 		internal const int DefaultNumberOnlineToLoad = 30;
 		public int numberOnlineToLoad = 0;
 
-		public override void UpdateUI(GameTime gameTime)
-		{
-			base.UpdateUI(gameTime);
-
-			if (Main.netMode == 1 && ModContent.GetInstance<CheatSheetServerConfig>().DisableCheatsForNonHostUsers && !IsPlayerLocalServerOwner(Main.LocalPlayer))
-				return;
-
-			if (PaintToolsEx.schematicsToLoad != null && numberOnlineToLoad > 0 && CheatSheet.instance.paintToolsUI.view.childrenToRemove.Count == 0)
-			{
-				PaintToolsEx.LoadSingleSchematic();
-				//CheatSheet.instance.paintToolsUI.view.ReorderSlots();
-			}
-
-			if (PaintToolsSlot.updateNeeded)
-			{
-				bool oneUpdated = false;
-				foreach (var item in paintToolsUI.view.slotList)
-				{
-					if (item.texture == TextureAssets.MagicPixel.Value)
-					{
-						item.texture = item.MakeThumbnail(item.stampInfo);
-						oneUpdated = true;
-						break;
-					}
-				}
-				if (!oneUpdated)
-					PaintToolsSlot.updateNeeded = false;
-			}
-		}
-
 		//public override void PostDrawFullscreenMap(ref string mouseText)
 		//{
 		//	Main.spriteBatch.DrawString(FontAssets.MouseText.Value, "Testing Testing", new Vector2(Main.screenWidth / 2, Main.screenHeight / 2), Color.Pink, 0.0f, new Vector2(), 0.8f, SpriteEffects.None, 0.0f);
 		//}
-
-		private int lastmode = -1;
-
-		public override void ModifyInterfaceLayers(List<GameInterfaceLayer> layers)
-		{
-			if (Main.netMode == 1 && ModContent.GetInstance<CheatSheetServerConfig>().DisableCheatsForNonHostUsers && !IsPlayerLocalServerOwner(Main.LocalPlayer))
-				return;
-
-			if (Main.netMode != lastmode)
-			{
-				lastmode = Main.netMode;
-				if (Main.netMode == 0)
-				{
-					SpawnRateMultiplier.HasPermission = true;
-					foreach (var key in herosPermissions.Keys.ToList()) {
-						herosPermissions[key] = true;
-					}
-				}
-				CheatSheet.instance.hotbar.ChangedConfiguration();
-			}
-			int MouseTextIndex = layers.FindIndex(layer => layer.Name.Equals("Vanilla: Mouse Text"));
-			if (MouseTextIndex != -1)
-			{
-				layers.Insert(MouseTextIndex, new LegacyGameInterfaceLayer(
-					"CheatSheet: All Cheat Sheet",
-					delegate
-					{
-						ModContent.GetInstance<AllItemsMenu>().DrawUpdateAll(Main.spriteBatch);
-						return true;
-					},
-					InterfaceScaleType.UI)
-				);
-
-				layers.Insert(MouseTextIndex, new LegacyGameInterfaceLayer(
-					"CheatSheet: Paint Tools",
-					delegate
-					{
-						ModContent.GetInstance<AllItemsMenu>().DrawUpdatePaintTools(Main.spriteBatch);
-						return true;
-					},
-					InterfaceScaleType.Game)
-				);
-			}
-
-			MouseTextIndex = layers.FindIndex(layer => layer.Name.Equals("Vanilla: Inventory"));
-			if (MouseTextIndex != -1)
-			{
-				layers.Insert(MouseTextIndex, new LegacyGameInterfaceLayer(
-					"CheatSheet: Extra Accessories",
-					delegate
-					{
-						ModContent.GetInstance<AllItemsMenu>().DrawUpdateExtraAccessories(Main.spriteBatch);
-						return true;
-					},
-					InterfaceScaleType.UI)
-				);
-			}
-		}
 
 		public static bool IsPlayerLocalServerOwner(Player player) {
 			if(Main.netMode == 1) {
